web/requirements.txt
# Web framework
fastapi==0.115.12
uvicorn==0.34.3

# Configuration
python-dotenv==1.1.0

# Database
sqlalchemy==2.0.48
pymysql==1.1.2

# Oracle (optional, install only when needed)
# oracledb==3.1.0
